домашнее задание по модулям
