+++ js/data.js
@@ -1,0 +1,35 @@
+const NAMES = [
+  'Артём',
+  'Саша',
+  'Настя',
+  'Стас',
+  'Марина',
+  'Ксюша',
+  'Вера',
+  'Полина',
+  'Игорь'
+];
+
+const SENTANCES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const DESCRIPTIONS = [
+  'описание фото',
+  'моя ласточка',
+  'Отдыхаю',
+  'Мой хозяин идиот',
+  'Ласточка деда',
+  'Закат',
+  'Якобы мой завтрак',
+  'Мой питомец',
+  'Скучно',
+  'Рассвет',
+  'Пришлось идти пешком...'
+];
+export {DESCRIPTIONS, SENTANCES, NAMES};
